unit test for loading from checkpoint
