+++ tests/test_torch_agent.py
@@ -6,81 +6,20 @@
 
 """Unit tests for TorchAgent."""
 
+import os
 import unittest
 from parlai.core.agents import Agent
+from parlai.core.testing_utils import capture_output, tempdir
 
 from collections import deque
 
 SKIP_TESTS = False
 try:
-    from parlai.core.torch_agent import TorchAgent, Output
+    from parlai.core.torch_agent import Output
+    from parlai.agents.test_agents.dummy_torch_agent import TorchAgent, MockDict
     import torch
 except ImportError:
     SKIP_TESTS = True
-
-
-class MockDict(Agent):
-    """Mock Dictionary Agent which just implements indexing and txt2vec."""
-
-    null_token = '__null__'
-    NULL_IDX = 0
-    start_token = '__start__'
-    BEG_IDX = 1001
-    end_token = '__end__'
-    END_IDX = 1002
-    p1_token = '__p1__'
-    P1_IDX = 2001
-    p2_token = '__p2__'
-    P2_IDX = 2002
-
-    def __init__(self, opt, shared=None):
-        """Initialize idx for incremental indexing."""
-        self.idx = 0
-
-    def __getitem__(self, key):
-        """Return index of special token or return the token."""
-        if key == self.null_token:
-            return self.NULL_IDX
-        elif key == self.start_token:
-            return self.BEG_IDX
-        elif key == self.end_token:
-            return self.END_IDX
-        elif key == self.p1_token:
-            return self.P1_IDX
-        elif key == self.p2_token:
-            return self.P2_IDX
-        else:
-            self.idx += 1
-            return self.idx
-
-    def __setitem__(self, key, value):
-        pass
-
-    def add_cmdline_args(self, *args, **kwargs):
-        """Add CLI args."""
-        pass
-
-    def txt2vec(self, txt):
-        """Return index of special tokens or range from 1 for each token."""
-        self.idx = 0
-        return [self[tok] for tok in txt.split()]
-
-
-class TorchAgent(TorchAgent):
-    """Use MockDict instead of regular DictionaryAgent."""
-
-    @staticmethod
-    def dictionary_class():
-        """Replace normal dictionary class with mock one."""
-        return MockDict
-
-    def train_step(self, batch):
-        """Return confirmation of training."""
-        return Output(['Training {}!'.format(i) for i in range(len(batch.text_vec))])
-
-    def eval_step(self, batch):
-        """Return confirmation of evaluation."""
-        return Output(['Evaluating {}!'.format(i) for i in range(len(batch.text_vec))])
 
 
 def get_agent(**kwargs):
@@ -103,11 +42,13 @@ def get_agent(**kwargs):
 class TestTorchAgent(unittest.TestCase):
     """Basic tests on the util functions in TorchAgent."""
 
+    @unittest.skipIf(SKIP_TESTS, "Torch not installed.")
     def test_mock(self):
         """Just make sure we can instantiate a mock agent."""
         agent = get_agent()
         self.assertTrue(isinstance(agent.dict, MockDict))
 
+    @unittest.skipIf(SKIP_TESTS, "Torch not installed.")
     def test_share(self):
         """Make sure share works and shares dictionary."""
         agent = get_agent()
@@ -617,6 +558,7 @@ class TestTorchAgent(unittest.TestCase):
             ['By Grabthar’s hammer.', 'By the suns of Worvan.'],
         )
 
+    @unittest.skipIf(SKIP_TESTS, "Torch not installed.")
     def test__add_person_tokens(self):
         """Make sure person tokens are added to the write place in text."""
         agent = get_agent()
@@ -633,6 +575,7 @@ class TestTorchAgent(unittest.TestCase):
         idx = text.rfind('\n') + 1
         self.assertEqual(out, text[:idx] + prefix + ' ' + text[idx:])
 
+    @unittest.skipIf(SKIP_TESTS, "Torch not installed.")
     def test_history(self):
         """Test different dialog history settings."""
         # try with unlimited history
@@ -769,6 +712,7 @@ class TestTorchAgent(unittest.TestCase):
         text = agent.history.get_history_str()
         self.assertEqual(text, 'I am Groot. Groot! I am Groot.')
 
+    @unittest.skipIf(SKIP_TESTS, "Torch not installed.")
     def test_last_reply(self):
         """Make sure last reply returns expected values."""
         agent = get_agent()
@@ -881,6 +825,58 @@ class TestTorchAgent(unittest.TestCase):
         reply = agent.batch_act(obs_elabs_vecs)
         for i in range(len(obs_elabs_vecs)):
             self.assertEqual(reply[i]['text'], 'Evaluating {}!'.format(i))
+
+    @unittest.skipIf(SKIP_TESTS, "Torch not installed.")
+    def test_resume_checkpoint(self):
+        """Make sure when resuming training that model uses appropriate mf.
+
+        Copy train_model from testing_utils to directly access agent.
+        """
+        import parlai.scripts.train_model as tms
+
+        def get_popt_and_tl(opt):
+            parser = tms.setup_args()
+            parser.set_params(**opt)
+            popt = parser.parse_args(print_args=False)
+            for k, v in opt.items():
+                popt[k] = v
+            return popt, tms.TrainLoop(popt)
+
+        def get_opt(init_mf, mf):
+            return {
+                'task': 'integration_tests',
+                'init_model': init_mf,
+                'model': 'parlai.agents.test_agents.dummy_torch_agent:TorchAgent',
+                'model_file': mf,
+                'num_epochs': 3,
+                'validation_every_n_epochs': 1,
+                'save_after_valid': True,
+                'log_every_n_secs': 10
+            }
+
+        with capture_output():
+            with tempdir() as tmpdir:
+                # First train model with init_model path set
+                mf = os.path.join(tmpdir, 'model')
+                init_mf = os.path.join(tmpdir, 'init_model')
+                with open(init_mf, 'w') as f:
+                    f.write(' ')
+                opt = get_opt(init_mf, mf)
+                popt, tl = get_popt_and_tl(opt)
+                agent = tl.agent
+                # init model file should be set appropriately
+                init_model_file, is_finetune = agent._get_init_model(popt, None)
+                self.assertEqual(init_model_file, init_mf)
+                self.assertTrue(is_finetune)
+                valid, test = tl.train()
+                # now, train the model for another epoch
+                opt = get_opt('{}.checkpoint'.format(mf), mf)
+                opt['load_from_checkpoint'] = True
+                popt, tl = get_popt_and_tl(opt)
+                agent = tl.agent
+                init_model_file, is_finetune = agent._get_init_model(popt, None)
+                self.assertEqual(init_model_file, '{}.checkpoint'.format(mf))
+                self.assertFalse(is_finetune)
 
 
 if __name__ == '__main__':
